Restructure airflow.timetables
